Rework Part 2
Removed the need for AceTimer library.

Finished up the Advanced Search frame rework.  It's not much faster!

Added checks to make sure that the player Guild doesn't get cached, which would result in inaccurate counts.

Reverted some Guild Bank changes slightly.  The old code is back but has been optimized.  I removed the queue system and just query all the tabs when the guild bank is opened.  The only difference is that after the initial scan, only the tabs that are visited (clicked on) will be queried updated in the database.  NOTE: There is a delay and a bug on the WOW client where queried tabs don't always return the correct item count until you actually visit the tab.  No way around this.