Rework Part 5
Reworked the Gold display.  It's now an independent scrolling frame instead of a tooltip.  The tooltip was causing issues with people that had a ridiculous amount of characters across multiple servers.  In addition it was distorting the tooltip and preventing it from being displayed properly.  I had to move it to it's own frame to fix this problem.

Whitelist Module has been reworked.