Rework Part 6
Removed the custom labels used by BagSync for AceGUI.

Reworked the Sort Order window and have added it to the slash commands and minimap. Reworded a few things related to it in the enUS localization.

Added Tooltip Cache resets for various windows and options that may affect the Tooltip display such as Blacklist, Whitelist, Profiles and Sort Order.

Reworked the Profiles window.