Rework Part 7
Work on the new Item Details window is moving along.  It will display a detailed summary of all characters and locations where a specific item is located.  In addition to providing added information.

Updated the new Gold Window to display only the gold values if a character has any gold on them.  If they don't they it will default back to silver and copper.

Small update to the Professions window to provide more information on a character in the tooltip to help distinguish between characters.

Reworked a bit of code to prevent some functions from being too spammy or at least being called WAY too much.