Rework Part 15
Item Cache is now generated when a user logins and checks when they open the Search frame.  This should definitely increase the speed in which searches are done.

Fixed an issue where the new Details window was not being elevated when displayed above other frames.

Updated the BagSync timer functions as it was deleting the previous run function before it was finished.  This uses a priority system now and will delete them as they performed.