Gold/Professions and Guild Bank Updates
Guild Bank money should now properly update when withdrawing or depositing gold.  This includes guild repairs outside of the guild bank.

Fixed an issue with the BagSync timer functions that was ignoring certain timers from repeating again.

Updated the Gold and Professions windows to display Realm Names if the BNET option is enabled.  In addition, I have moved the tags to the end of the character names to ensure they are always shown.

Fixed a minor scroll issue with the Gold Window.

Fixed a slight issue in the Currency tooltip display that was adding a space at the top of the currency name.